doesnt optimize if file already exists

+++ src/MarketTickFileOptimizer.java
@@ -10,6 +10,10 @@ import java.util.StringTokenizer;
 public class MarketTickFileOptimizer {
 	
 	private static void work(File file, String outSignature, boolean autoSignature) {
+		if (file.exists()) {
+			System.out.println("file already exists. fuhgeddaboudit");
+			return;
+		}
 		try (FileInputStream fis = new FileInputStream(file);
 				BufferedReader br = new BufferedReader(new InputStreamReader(fis));
 				FileOutputStream fos = new FileOutputStream(new File(file.getAbsolutePath().substring(0, file.getAbsolutePath().lastIndexOf('.')) + "_Optimized.csv"), true)) {
